Merged PR 1048435: Defining a new property for AST.XML for connecting a single item to a collection
Sometimes we need to change an existing SQL object property, which we already have support in Scriptdom, from one item property to a collection. This PR introduces a new property for AST file for adding a collection and connecting it to the existing property. This is to make sure we don't introduce breaking changes in. So using this new feature, we add a new collection and keep the current property but set CollectionFirstItem to true so in the generated file it gets and sets the first value of the collection.

This is how generated code will look like using the new xml property:

`		private List<Identifier> _distributionColumns = new List<Identifier>();

		/// <summary>
		/// The column reference in the HASH option for a table distribution policy.
		/// </summary>
		public IList<Identifier> DistributionColumns
		{
			get { return _distributionColumns; }
		}

		/// <summary>
		/// The column reference in the HASH option for a table distribution policy.
		/// </summary>
		public Identifier DistributionColumn
		{
			get { return DistributionColumns.Count > 0 ? DistributionColumns[0] : null; }
			set
			{
				if (DistributionColumns.Count == 0)
				{
					DistributionColumns.Add(value);
				}
				else
				{
					DistributionColumns[0] = value;
				}
			}
		}`