Merged PR 1274538: fixing the build warnings
fixing the build warnings

----
#### AI description (iteration 1)
#### PR Classification
This pull request is primarily a code cleanup, aimed at fixing build warnings in the project.

#### PR Summary
The changes in this pull request address build warnings by modifying the code in two files. The modifications improve the code quality and maintainability, and ensure a smoother build process.
- In the file `TSqlParserTest.cs`, the method of creating a parser has been changed from `parser.Create(version, false)` to `TSqlParser.CreateParser(version, false)`.
- In the file `TSql160.g`, the boolean variable `vHasColumnList` and its related code have been removed, as it was not being used, thus eliminating unnecessary code.