Merged PR 1367589: handle json data type while parsing
Json data type has been added to SqlDataTypeOption. However, Json data type is not mapped to this enum while parsing t-sql scripts. For example, for script Create Table t1(c1 json), type of column c1 will be mapped as a UserDefinedType.
This PR fixed this issue by adding a ParseDataType160 method in TSql160ParserBaseInternal.cs and map token "json" to SqlDataTypeOption.Json.
A test case is added in CreateTableTests160.sql to cover this.

----
#### AI description  (iteration 1)
#### PR Classification
This pull request introduces a new feature to handle JSON data type while parsing in the SQL script.

#### PR Summary
The changes in this pull request allow the SQL parser to recognize and handle JSON data types. This enhancement improves the versatility of the parser and expands its compatibility with different data types.
- In `TSql160ParserBaseInternal.cs`, a new method `ParseDataType160` was added to handle the parsing of JSON data type.
- In `TSql160.g`, the method call was updated from `ParseDataType100` to `ParseDataType160` to utilize the new JSON parsing functionality.
- Test scripts in `CreateTableTests160.sql` and `TestScripts/CreateTableTests160.sql` were updated to include tests for tables with JSON columns, ensuring the new feature works as expected.

Related work items: #3264300