Merged PR 1381485: Adding release notes for release 9123
Adding release notes for release 9123

----
#### AI description  (iteration 1)
#### PR Classification
This pull request is primarily for documentation, providing release notes for the new release 9123 of Microsoft.SqlServer.TransactSql.ScriptDom.

#### PR Summary
The pull request adds a new markdown file for the release notes of version 161.9123.0 of Microsoft.SqlServer.TransactSql.ScriptDom. The notes detail the target platform support, dependencies, fixes, and changes in this release.
- Added a new file `/release-notes/161.91/161.9123.0.md` detailing the release notes for version 161.9123.0, including the support for .NET Framework 4.6.2, .NET 6, and .NET Standard 2.0+, as well as a fix for parsing JSON column type in a table.