Merged PR 1450917: Adding CTAS support to the 160Parser
This change adds the ability to parse the CTAS statement which has the "Create Table TblName AS Select" format. So far we support "Create Table tableName With(Distribution) AS Select...) syntax only. This change fixes the syntax error at Select and allows the DW users to run the sql queries with the syntax.

----
#### AI description  (iteration 1)
#### PR Classification
New feature: Adding support for CTAS (Create Table As Select) statements in the 160Parser.

#### PR Summary
This pull request introduces support for parsing CTAS statements in the TSql160Parser and includes corresponding unit tests.
- Added `CreateCTASParser160Test` method in `Test/SqlDom/TSqlParserTest.cs` to validate CTAS statement parsing.
- Updated `TSql160.g` to handle `As` keyword for CTAS statements.
- Minor import addition in `Test/SqlDom/PhaseOneParserTest.cs`.